submodules/make-sbom/requirements.txt
python-box==7.1.1
spdx-tools==0.8.2
